refactor: extract PlaceResolver from LocationManager
Lifts findOrCreate, nearestExisting, and private POI/geocoding helpers
into a standalone PlaceResolver enum so QuickCaptureService can reuse
them without depending on LocationManager.
